separated the impl header from the cpp file
- tests are passing
- openmc is compiling (pimpl structure is working)
