+++ test/CMakeLists.txt
@@ -26,7 +26,7 @@ catch_discover_tests(test_hostwrite_initialization)
 
 add_executable(test_pumi_tally_impl_methods test_pumi_tally_impl_methods.cpp)
 target_include_directories(test_pumi_tally_impl_methods PRIVATE ${PROJECT_SOURCE_DIR}/src/pumitallyopenmc)
-target_link_libraries(test_pumi_tally_impl_methods PUBLIC Catch2::Catch2WithMain pumipic::pumipic)
+target_link_libraries(test_pumi_tally_impl_methods PUBLIC Catch2::Catch2WithMain pumitallyopenmc pumipic::pumipic)
 catch_discover_tests(test_pumi_tally_impl_methods)
 
 if(PUMI_ENABLE_DEGAS2)
