+++ test/test_pumi_tally_impl_methods.cpp
@@ -1,6 +1,7 @@
 //
 // Created by Fuad Hasan on 2/3/25.
 //
+#include "pumitally_impl.hpp"
 #include <Omega_h_build.hpp>
 #include <Omega_h_library.hpp>
 #include <Omega_h_vtk.hpp>
@@ -11,13 +12,9 @@
 // - Sections are not working for some reason. Saying using MPI functions before
 // or after MPI init or finalize
 // - sections are marked with comments for now
-// - TODO Remove including cpp by creating another internal header
 // - Look at this gist to verify this in python:
 // https://gist.github.com/Fuad-HH/5e0aed99f271617e283e9108091fb1cb
 // *****************************************************************************************************//
-
-// TODO: Remove it by having another header file
-#include "pumipic_particle_data_structure.cpp"
 
 bool is_close(const double a, const double b, double tol = 1e-8) {
   return std::abs(a - b) < tol;
